third_party/WebKit/LayoutTests/editing/undo/undo-forward-delete-expected.txt

mac
PASS selection.anchorNode.data is "This word should be selected only on mac."
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This word should be selected only on mac."
PASS selection.focusOffset is 9
PASS selection.toString() is "word"
win
PASS selection.anchorNode.data is "This word should be selected only on mac."
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This word should be selected only on mac."
PASS selection.focusOffset is 5
PASS selection.toString() is ""
unix
PASS selection.anchorNode.data is "This word should be selected only on mac."
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This word should be selected only on mac."
PASS selection.focusOffset is 5
PASS selection.toString() is ""
android
PASS selection.anchorNode.data is "This word should be selected only on mac."
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This word should be selected only on mac."
PASS selection.focusOffset is 5
PASS selection.toString() is ""
PASS successfullyParsed is true

TEST COMPLETE
